Axe Code Continued
The axe code now checks if the player meets the requirements to craft an axe, and denies them if they do not.

The axe code now removes the materials when successfully crafted, and if somehow the values dip below 0, sets values to 0.

The code prevents the same axe from being crafted twice.